requirements.txt
discord.py
python-dotenv
oauth2client
gspread